Make useReview hook return useful review stats, apply it in neccesary places, replace use of reputation with review stats
