+++ website/src/hooks/subsquid/useReviews.tsx
@@ -23,11 +23,50 @@ export default function useReviews(
     skip: !targetAddress,
   });
 
-  return useMemo(
-    () => ({
-      data: data ? (data?.reviews as ExtendedReview[]) : undefined,
+  return useMemo(() => {
+    const reviews = data ? (data.reviews as ExtendedReview[]) : undefined;
+
+    // Compute review statistics
+    const stats = (() => {
+      if (!reviews || reviews.length === 0) {
+        return {
+          totalReviews: 0,
+          positiveReviews: 0,
+          negativeReviews: 0,
+          positiveReviewPercentage: 0,
+          actualAverageRating: 0,
+        };
+      }
+
+      type Acc = { total: number; positive: number; negative: number; sum: number };
+
+      const acc = reviews.reduce(
+        (acc: Acc, review) => {
+          acc.total++;
+          acc.sum += review.rating ?? 0;
+          if ((review.rating ?? 0) >= 3) acc.positive++;
+          else acc.negative++;
+          return acc;
+        },
+        { total: 0, positive: 0, negative: 0, sum: 0 }
+      );
+
+      const avg = acc.total > 0 ? acc.sum / acc.total : 0;
+      const percentage = acc.total > 0 ? Math.round((acc.positive / acc.total) * 100) : 0;
+
+      return {
+        totalReviews: acc.total,
+        positiveReviews: acc.positive,
+        negativeReviews: acc.negative,
+        positiveReviewPercentage: percentage,
+        actualAverageRating: avg,
+      };
+    })();
+
+    return {
+      data: reviews,
       ...rest,
-    }),
-    [targetAddress, offset, limit, data, rest]
-  );
+      ...stats,
+    };
+  }, [targetAddress, offset, limit, data, rest]);
 }
